pyproject.toml
[tool.poetry]
name = "shamusic"
version = "0.1.0"
description = "APP the songs"
authors = ["AlekysCoelho <alekyscoelho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.2"
django-filter = "^24.2"
pillow = "^10.4.0"
python-magic = "^0.4.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
